+++ server/db/data.js
@@ -1,0 +1,275 @@
+
+var data = {
+    user: [
+        {
+            name: "Annelise Thorsen",
+            img: '/images/annelise.jpg',
+            email: "thorsen120@gmail.com",
+            password: "abc123",
+            rating: "5"
+        },
+        {
+            name: "Ella Pitassi",
+            img: '/images/ella.jpg',
+            email: "ellapitassi23@gmail.com",
+            password: "abc123",
+            rating: "5"
+        },
+        {
+            name: "Jennifer Neale",
+            img: '/images/jennifer.jpg',
+            email: "jeneale6@gmail.com",
+            password: "abc123",
+            rating: "5"
+        },
+        {
+            name: "Guang Zhu",
+            img: '/images/guang.jpg',
+            email: "bha.guang@gmail.com",
+            password: "abc123",
+            rating: "5"
+        },
+        {
+            name: "Anule Ndukwu",
+            img: '/images/anule.jpg',
+            email: "anule@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "April Rueb",
+            img: '/images/april.jpg',
+            email: "april@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Arianna Lanz",
+            img: '/images/arianna.jpg',
+            email: "arianna@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Ashi",
+            img: '/images/ashi.jpg',
+            email: "ashi@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Emily Intersimone",
+            img: '/images/emilyi.jpg',
+            email: "emilyi@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Blanca Sanchez",
+            img: '/images/blanca.jpg',
+            email: "blanca@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Cara Lang",
+            img: '/images/cara.jpg',
+            email: "cara@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Caryn McCarthy",
+            img: '/images/caryn.jpg',
+            email: "caryn@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Cecilia Chang",
+            img: '/images/cecilia.jpg',
+            email: "ceciliajaimie@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Christina Ng",
+            img: '/images/christina.jpg',
+            email: "christina@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Evlis Henry",
+            img: '/images/evlis.jpg',
+            email: "evlis@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Emily Jordan",
+            img: '/images/emily.jpg',
+            email: "emilyj@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Erica Chai",
+            img: '/images/erica.jpg',
+            email: "erica@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Joyce Ren",
+            img: '/images/joyce.jpg',
+            email: "joyce@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Keri Miller",
+            img: '/images/keri.jpg',
+            email: "keri@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Layla Hedges",
+            img: '/images/layla.jpg',
+            email: "layla@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Lina Morales",
+            img: '/images/lina.jpg',
+            email: "lina@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Maria Betances",
+            img: '/images/maria.jpg',
+            email: "maria@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Moyouri Bhattacharjee",
+            img: '/images/momo.jpg',
+            email: "momo@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Renee Sajedian",
+            img: '/images/renee.jpg',
+            email: "renee@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Samantha Zhang",
+            img: '/images/sam.jpg',
+            email: "sam@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Shannon Kendall",
+            img: '/images/shannon.jpg',
+            email: "shannon@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Shelby Rackley",
+            img: '/images/shelby.jpg',
+            email: "shelby@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Shiratie Prodhan",
+            img: '/images/shiratie.jpg',
+            email: "shiratie@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Priya Vaidyanath",
+            img: '/images/priya.jpg',
+            email: "priya@gmail.com",
+            password: "abc123",
+            rating: "4"
+        },
+        {
+            name: "Eleni Arvanitis",
+            img: '/images/eleni.jpg',
+            email: "eleni@gmail.com",
+            password: "abc123",
+            rating: "4"
+        }
+    ],
+    teachables: [
+        {
+            name: "React",
+            price: 10.00
+        },
+        {
+            name: "Redux",
+            price: 8.00
+        },
+        {
+            name: "Express",
+            price: 7.00
+        },
+        {
+            name: "Sequelize",
+            price: 5.00
+        },
+        {
+            name: "Crafting",
+            price: 3.00
+        },
+        {
+            name: "Knitting",
+            price: 4.00
+        },
+        {
+            name: "Writing",
+            price: 9.00
+        },
+        {
+            name: "Reading",
+            price: 10.00
+        },
+        {
+            name: "Numbers",
+            price: 11.00
+        },
+        {
+            name: "Eating",
+            price: 7.00
+        },
+        {
+            name: "Gaming",
+            price: 8.00
+        },
+        {
+            name: "FullStack",
+            price: 12.00
+        },
+        {
+            name: "ReactRedux",
+            price: 8.00
+        },
+        {
+            name: "Feminism",
+            price: 5.00
+        },
+    ]
+}
+
+module.exports = data
